refactor: GrimityDrawer를 공통 내비게이션 위젯으로 이동
