+++ lib/presentation/chat_message/chat_message_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/widgets.dart';
 import 'package:grimity/presentation/chat_message/chat_message_view.dart';
 import 'package:grimity/presentation/chat_message/view/chat_message_app_bar.dart';
 import 'package:grimity/presentation/chat_message/view/chat_message_field.dart';
-import 'package:grimity/presentation/drawer/grimity_drawer.dart';
+import 'package:grimity/presentation/common/widget/navigation/grimity_drawer.dart';
 
 class ChatMessagePage extends StatelessWidget {
   const ChatMessagePage({super.key, required this.chatId});
